+++ src/App.jsx
@@ -1,10 +1,10 @@
 import './App.css';
-// import CardListPage from './components/CardListPage';
-import CardAddPage from './components/CardAddPage';
+import CardListPage from './components/CardListPage/CardListPage';
+// import CardAddPage from './components/CardAddPage/CardAddPage';
 
 function App() {
-  // return <CardListPage />;
-  return <CardAddPage />;
+  return <CardListPage />;
+  // return <CardAddPage />;
 }
 
 export default App;
