--- src/components/CardListPage.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import Card from './Card';
-
-function CardListPage() {
-  return (
-    <div className="app">
-      <Card />
-    </div>
-  );
-}
-
-export default CardListPage;
