--- src/components/PageTitle.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function PageTitle(props) {
-  return <h2 className="page-title">카드 추가</h2>;
-}
-
-export default PageTitle;
