+++ src/components/common/PageTitle.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function PageTitle({ title }) {
+  return <h2 className="page-title">{title}</h2>;
+}
+
+export default PageTitle;
